Key Display In Progress

+++ src/main/java/com/edufyy/backend/service/ProficiencyService.java
@@ -26,14 +26,14 @@ public class ProficiencyService {
 
         // TODO: Proficiency Calculation Logic
         // Map difficulty to answers
-        Map<Integer, List<Answer>> mapDifficultyToAnswers = mapDifficultyToAnswers(answers);
+        Map<Double, List<Answer>> mapDifficultyToAnswers = mapDifficultyToAnswers(answers);
         // Count total answers
         Integer totalAnswers = answers.size();
 
         // TODO: Find individual proficiency and calculate key proficiency
         // Find subproficiency per question difficulty
         List<Double> subProficiency = new ArrayList<>();
-        for (Map.Entry<Integer, List<Answer>> difficultyToAnswerEntry : mapDifficultyToAnswers.entrySet()) {
+        for (Map.Entry<Double, List<Answer>> difficultyToAnswerEntry : mapDifficultyToAnswers.entrySet()) {
             Integer totalRight = 0;
             for (Answer answer : difficultyToAnswerEntry.getValue())
                 if (questionService.findByAnswer(answer.getQuestionKey(), answer.getQuestionNumber(), answer.getAnswer()) != null)
@@ -53,12 +53,12 @@ public class ProficiencyService {
         return response;
     }
 
-    public Map<Integer, List<Answer>> mapDifficultyToAnswers(List<Answer> answers) {
-        Map<Integer, List<Answer>> mapDifficultyToAnswers = new HashMap<>();
+    public Map<Double, List<Answer>> mapDifficultyToAnswers(List<Answer> answers) {
+        Map<Double, List<Answer>> mapDifficultyToAnswers = new HashMap<>();
 
         for (Answer answer : answers) {
             Question question = questionService.findQuestionByAnswer(answer.getQuestionKey(), answer.getQuestionNumber());
-            Integer answerDifficulty = question.getStudentDiffuculty();
+            Double answerDifficulty = question.getStudentDiffuculty();
             List<Answer> tempAnswers;
             if (mapDifficultyToAnswers.get(answerDifficulty) == null)
                 tempAnswers = new ArrayList<>();
